+++ Math/Tensor.h
@@ -1,20 +1,22 @@
 #pragma once
 #ifndef NEURALNETWORK_TENSOR_H
 #define NEURALNETWORK_TENSOR_H
+#include <stddef.h>
 
 
 typedef struct {
     int nDim;
-    int *shape;
-    int *stride;
+    size_t *shape;
+
+    size_t *stride;
     float *data;
-    long total;
+    size_t total;
     int total_valid;
     // for transposition if 1 it owns data
     int isOwner;
 }Tensor;
 
-Tensor *createTensor(int ndim, const int *shape);
+Tensor *createTensor(int ndim, const size_t *shape);
 void deleteTensor(Tensor *tensor);
 int equals(Tensor *tensorA, Tensor *tensorB);
 float getValue(Tensor *tensor, const int *indices);
@@ -25,9 +27,9 @@ void overwriteTensor(Tensor *tensor, const float *newValues);
 void printTensorShape(Tensor *tensor);
 void printTensorSize(Tensor *tensor);
 void printTensorDimension(Tensor *tensor);
-void printTensorRecursive(const Tensor *tensor, int dim, int offset);
+void printTensorRecursive(const Tensor *tensor, int dim, size_t offset);
 void printTensor(const Tensor *tensor);
-void printTensorHeadRecursive(const Tensor *tensor, int dim, int offset, int limit);
+void printTensorHeadRecursive(const Tensor *tensor, int dim, size_t offset, int limit);
 void printTensorHead(const Tensor *tensor, int limit);
 // Tensor *Multiply(Tensor *tensorA, Tensor *tensorB);
 int findContractableDims(const Tensor *A, const Tensor *B,
